+++ lib/ferrum/browser.rb
@@ -49,6 +49,9 @@ module Ferrum
     # @option options [Boolean] :incognito (true)
     #   Create an incognito profile for the browser startup window.
     #
+    # @option options [Boolean] :dockerize (false)
+    #   Add CLI flags to a browser to run in a container.
+    #
     # @option options [Boolean] :xvfb (false)
     #   Run browser in a virtual framebuffer.
     #