+++ lib/ferrum/browser/options.rb
@@ -14,7 +14,7 @@ module Ferrum
       attr_reader :window_size, :logger, :ws_max_receive_size,
                   :js_errors, :base_url, :slowmo, :pending_connection_errors,
                   :url, :ws_url, :env, :process_timeout, :browser_name, :browser_path,
-                  :save_path, :proxy, :port, :host, :headless, :incognito, :browser_options,
+                  :save_path, :proxy, :port, :host, :headless, :incognito, :dockerize, :browser_options,
                   :ignore_default_browser_options, :xvfb, :flatten
       attr_accessor :timeout, :default_user_agent
 
@@ -28,6 +28,7 @@ module Ferrum
         @js_errors = @options.fetch(:js_errors, false)
         @headless = @options.fetch(:headless, true)
         @incognito = @options.fetch(:incognito, true)
+        @dockerize = @options.fetch(:dockerize, false)
         @flatten = @options.fetch(:flatten, true)
         @pending_connection_errors = @options.fetch(:pending_connection_errors, true)
         @process_timeout = @options.fetch(:process_timeout, PROCESS_TIMEOUT)