+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - `Ferrum::Network::Response#body!` returns body or throws error if implicable
+- `Ferrum::Browser#new(dockerize: true)` whether to add CLI flags to run a browser in a container, `false` by default.
 
 ### Changed
 - `Ferrum::Network::Response#body` returns body or nil in case of errors